Put Create Facility so that it prompts when click Manage Events for the first time
