+++ app/com/example/intels_app/MainPageActivity.html
@@ -79,7 +79,7 @@ loadScripts(document, 'script');</script>
 <div class="inheritance">androidx.activity.ComponentActivity
 <div class="inheritance">androidx.fragment.app.FragmentActivity
 <div class="inheritance">androidx.appcompat.app.AppCompatActivity
-<div class="inheritance">com.example.intels_app.MainPageActivity</div>
+<div class="inheritance">com.example.intels_app.MainActivity</div>
 </div>
 </div>
 </div>
