+++ app/com/example/intels_app/MainPageActivityTest.html
@@ -71,7 +71,7 @@ loadScripts(document, 'script');</script>
 <h1 title="Class MainPageActivityTest" class="title">Class MainPageActivityTest</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/17/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
-<div class="inheritance">com.example.intels_app.MainPageActivityTest</div>
+<div class="inheritance">com.example.intels_app.MainActivityTest</div>
 </div>
 <section class="class-description" id="class-description">
 <hr>
