+++ app/src/androidTest/java/com/example/intels_app/AdminEventsTest.java
@@ -8,17 +8,10 @@ import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
 
-import static java.util.regex.Pattern.matches;
-
-import android.view.View;
-
-import androidx.core.content.ContextCompat;
-import androidx.test.espresso.ViewAssertion;
 import androidx.test.espresso.intent.Intents;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
-import org.hamcrest.Matcher;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
@@ -50,7 +43,7 @@ public class AdminEventsTest {
     @Test
     public void testBackButton() {
         onView(withId(R.id.back_button)).perform(click());
-        intended(hasComponent(MainPageActivity.class.getName()));
+        intended(hasComponent(MainActivity.class.getName()));
     }
 
 }
