+++ app/src/androidTest/java/com/example/intels_app/ManageEventsActivityTest.java
@@ -2,7 +2,6 @@ package com.example.intels_app;
 
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.click;
-import static androidx.test.espresso.action.ViewActions.typeText;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
@@ -19,7 +18,7 @@ import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import androidx.test.ext.junit.rules.ActivityScenarioRule;
+
 import androidx.test.filters.LargeTest;
 
 @RunWith(AndroidJUnit4.class)
@@ -40,7 +39,7 @@ public class ManageEventsActivityTest {
     @Test
     public void testBackButton() {
         onView(withId(R.id.backButton)).perform(click());
-        intended(hasComponent(MainPageActivity.class.getName()));
+        intended(hasComponent(MainActivity.class.getName()));
     }
 
     @Test
