+++ app/src/androidTest/java/com/example/intels_app/ManageFacilityTest.java
@@ -1,7 +1,6 @@
 package com.example.intels_app;
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.click;
-import static androidx.test.espresso.action.ViewActions.typeText;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
@@ -9,12 +8,7 @@ import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.intent.Intents.intended;
 import static androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent;
 
-import android.net.Uri;
-import android.view.View;
-
-import androidx.test.core.app.ActivityScenario;
 import androidx.test.espresso.intent.Intents;
-import androidx.test.espresso.intent.rule.IntentsTestRule;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 
@@ -51,7 +45,7 @@ public class ManageFacilityTest {
     @Test
     public void testbackButton(){
         onView(withId(R.id.back_button)).perform(click());
-        intended(hasComponent(MainPageActivity.class.getName()));
+        intended(hasComponent(MainActivity.class.getName()));
     }
 
 }
