+++ app/src/main/AndroidManifest.xml
@@ -54,7 +54,6 @@
         <activity android:name=".SelectRoleActivity" />
         <activity android:name=".ManageFacility" />
         <activity android:name=".CreateFacility" />
-        <activity android:name=".MainPageActivity" />
         <activity android:name=".SignUp" />
         <activity android:name=".SuccessWaitlistJoin" />
         <activity android:name=".DrawCompleteActivity"/>
