+++ app/src/main/java/com/example/intels_app/CreateFacility.java
@@ -1,7 +1,7 @@
 /**
  * When the user uses the app for the first time (new device ID), prompt them to create a facility profile
  * @author Janan Panchal
- * @see com.example.intels_app.MainPageActivity Creating a profile leads to main page
+ * @see com.example.intels_app.MainActivity Creating a profile leads to main page
  * @see com.example.intels_app.Facility Facility object
  */
 package com.example.intels_app;
@@ -18,7 +18,6 @@ import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ImageButton;
 import android.widget.ImageView;
 import android.widget.Toast;
 
@@ -28,10 +27,7 @@ import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.bumptech.glide.Glide;
-import com.example.intels_app.Facility;
-import com.example.intels_app.MainPageActivity;
 import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.installations.FirebaseInstallations;
 import com.google.firebase.messaging.FirebaseMessaging;
 import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
@@ -60,6 +56,9 @@ public class CreateFacility extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.create_facility);
 
+        deviceId = getIntent().getStringExtra("deviceId");
+
+        /*
         // Get the current user's device Id to add as a parameter to their facility profile
         FirebaseMessaging.getInstance().getToken()
                 .addOnCompleteListener(task -> {
@@ -70,7 +69,7 @@ public class CreateFacility extends AppCompatActivity {
                     } else {
                         Log.e("DeviceID", "Failed to get Firebase Instance ID", task.getException());
                     }
-                });
+                });*/
 
         imageView = findViewById(R.id.pfpPlaceholder);
 
@@ -161,7 +160,7 @@ public class CreateFacility extends AppCompatActivity {
         FirebaseFirestore.getInstance().collection("facilities").document(facilityName)
                 .set(facility)
                 .addOnSuccessListener(documentReference -> {
-                    Intent intent = new Intent(CreateFacility.this, MainPageActivity.class);
+                    Intent intent = new Intent(CreateFacility.this, ManageEventsActivity.class);
                     startActivity(intent);
                     finish();
                 })
