+++ app/src/main/java/com/example/intels_app/MainActivity.java
@@ -1,119 +1,134 @@
-/**
- * When the user uses the app for the first time (new device ID), prompt them to create a facility profile
- * If they are opening the as an existing user (registered device ID), open the app's main page
- */
 package com.example.intels_app;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.util.Log;
-import android.widget.Toast;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.navigation.ui.AppBarConfiguration;
 
-import com.example.intels_app.CreateFacility;
-import com.example.intels_app.MainPageActivity;
+import android.view.MenuInflater;
+import android.view.View;
+
+//import com.example.intels_app.databinding.ActivityMainBinding;
+import com.google.firebase.FirebaseApp;
+//import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.messaging.FirebaseMessaging;
+import com.google.firebase.storage.FirebaseStorage;
+import com.google.firebase.storage.StorageReference;
+
+import android.view.MenuItem;
+import android.widget.ImageButton;
+
+import android.widget.PopupMenu;
 
 /**
- * Main activity that initializes the app and determines if a user profile exists.
- * This activity retrieves the device ID using Firebase Messaging and checks
- * Firebase Firestore to see if a profile is associated with the device.
- * If a profile exists, the user is directed to the main page; otherwise,
- * they are redirected to the profile creation screen.
+ * MainActivity serves as the main dashboard of the application, providing navigation to various
+ * sections including event management, joining events, and administrative functions. It also includes
+ * a popup menu for additional options like viewing notifications and accessing user settings.
+ *
+ * This activity initializes Firebase and sets up click listeners for key buttons that navigate to
+ * different parts of the app.
+ * Author: Dhanshri Patel, Janan Panchal
  */
 public class MainActivity extends AppCompatActivity {
+    private FirebaseFirestore db;
+    //private FirebaseAuth fAuth;
+    private AppBarConfiguration appBarConfiguration;
+    private StorageReference storageRef;
+    //private ActivityMainBinding binding;
+    private QRCodeScanner qrCodeScanner;
 
-    /**
-     * Called when the activity is first created.
-     * Attempts to retrieve the Firebase Device ID, and uses it to
-     * check if the user's profile exists.
-     *
-     * @param savedInstanceState If the activity is being re-initialized after
-     *                           previously being shut down, this Bundle contains the most recent data.
-     */
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        FirebaseApp.initializeApp(this);
 
-        // Get Firebase Device ID
-        FirebaseMessaging.getInstance().getToken()
-                .addOnCompleteListener(task -> {
-                    if (task.isSuccessful() && task.getResult() != null) {
-                        String deviceId = task.getResult();
-                        checkUserExists(deviceId);
-                    } else {
-                        Log.e("DeviceID", "Failed to get Firebase Instance ID", task.getException());
-                        Toast.makeText(this, "Error retrieving device ID. Please try again.", Toast.LENGTH_LONG).show();
-                    }
-                });
-    }
+        db = FirebaseFirestore.getInstance();
+        //fAuth = FirebaseAuth.getInstance();
+        storageRef = FirebaseStorage.getInstance().getReference();
 
-    /**
-     * Checks whether a user profile exists for the given device ID.
-     * If a user profile exists, proceeds to the main functionality.
-     * If not, redirects the user to the profile creation screen.
-     *
-     * @param deviceId The unique device ID retrieved from Firebase.
-     */
-    private void checkUserExists(String deviceId) {
-        FirebaseFirestore.getInstance().collection("facilities")
-                .whereEqualTo("deviceId", deviceId)
-                .get()
-                .addOnCompleteListener(task -> {
-                    if (task.isSuccessful() && !task.getResult().isEmpty()) {
-                        // User exists, proceed to the main page
-                        proceedToApp();
-                    } else {
-                        // User does not exist, redirect to profile creation
-                        // redirectToCreateEntrantProfile(deviceId); // Uncomment line to implement
-                        redirectToCreateOrganizerProfile(deviceId);
-                    }
-                })
-                .addOnFailureListener(e -> {
-                    Log.e("Firestore", "Error checking user existence", e);
-                    Toast.makeText(this, "Error accessing user information. Please try again.", Toast.LENGTH_LONG).show();
-                });
-    }
+        setContentView(R.layout.main_page);
+        qrCodeScanner = new QRCodeScanner(this);
 
-    /**
-     * Proceeds to the main application functionality if a user profile exists.
-     * Starts {@link MainPageActivity} and finishes this activity.
-     */
-    private void proceedToApp() {
-        Intent intent = new Intent(MainActivity.this, MainPageActivity.class);
-        startActivity(intent);
-        finish(); // Close MainActivity
-    }
+        ImageButton optionsButton = findViewById(R.id.imageButton8);
+        optionsButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                showPopupMenu(v);
+            }
+        });
 
-    /**
-     * Redirects the user to the organizer profile creation activity if they are new.
-     * Passes the device ID to the profile creation activity.
-     *
-     * @param deviceId The unique device ID to be used in {@link CreateFacility}.
-     */
-    // Method to redirect to create profile if user is new
-    private void redirectToCreateOrganizerProfile(String deviceId) {
-        Intent intent = new Intent(MainActivity.this, CreateFacility.class); //Changed this line make it back to CreateFacility
-        intent.putExtra("deviceId", deviceId); // Pass device ID if needed in CreateProfileActivity
-        startActivity(intent);
-        finish(); // Close MainActivity
+        // Change it to qr code scanner button ion UI later
+        //qrCodeScanner.startScan();
+        ImageButton ViewWaitListButton = findViewById(R.id.imageButton7);
+        ViewWaitListButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(MainActivity.this, EventGridEntrantActivity.class);
+                startActivity(intent);
+            }
+        });
+
+        // Set up the Join Events button to navigate to ScanQRActivity
+        ImageButton joinEventButton = findViewById(R.id.joinEventButton);
+        joinEventButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(MainActivity.this, ScanQRActivity.class); //Change back to ScanQRActivity
+                startActivity(intent);
+            }
+        });
+
+        ImageButton manageApp = findViewById(R.id.manageAppButton);
+        manageApp.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(MainActivity.this, AdminLogin.class);
+                startActivity(intent);
+            }
+        });
+
+        ImageButton manageEventsButton = findViewById(R.id.manageEventsButton);
+        manageEventsButton.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(MainActivity.this, ManageEventsActivity.class);
+                startActivity(intent);
+            }
+        });
+
     }
-  
-    /**
-     * Redirects the user to the entrant profile creation activity if they are new.
-     * Passes the device ID to the profile creation activity.
-     * This method is currently commented out for future implementation.
-     *
-     * @param deviceId
-     */
-    private void redirectToCreateEntrantProfile(String deviceId) {
-        // Redirect to sign up page
-        // Intent intent = new Intent(MainActivity.this, CreateEntrantProfile.class);
-        // intent.putExtra("deviceId", deviceId);
-        // startActivity(intent);
-        // finish(); // Close MainActivity
+    // Method to display the popup menu
+    private void showPopupMenu(View view) {
+        PopupMenu popupMenu = new PopupMenu(this, view);
+        MenuInflater inflater = popupMenu.getMenuInflater();
+        inflater.inflate(R.menu.menu_main, popupMenu.getMenu());
+
+        // Handle menu item clicks
+        popupMenu.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
+            @Override
+            public boolean onMenuItemClick(MenuItem item) {
+                if (item.getItemId() == R.id.action_show_notifications) {
+                    // Navigate to NotificationActivity
+                    Intent intent = new Intent(MainActivity.this, NotificationActivity.class);
+                    startActivity(intent);
+                    return true;
+                } else if (item.getItemId() == R.id.action_settings) {
+                    Intent intent = new Intent(MainActivity.this, EditProfileActivity.class);
+                    startActivity(intent);
+                    return true;
+                }
+                return false;
+            }
+        });
+        // Show the popup menu
+        popupMenu.show();
+    }
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+        if (qrCodeScanner != null) {
+            qrCodeScanner.handleActivityResult(requestCode, resultCode, data);
+        }
     }
 }
