+++ app/src/main/java/com/example/intels_app/ManageFacility.java
@@ -30,7 +30,6 @@ import com.bumptech.glide.Glide;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.Query;
 import com.google.firebase.firestore.QuerySnapshot;
 import com.google.firebase.installations.FirebaseInstallations;
 import com.google.firebase.messaging.FirebaseMessaging;
@@ -107,7 +106,7 @@ public class ManageFacility extends AppCompatActivity {
         backButton.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(ManageFacility.this, MainPageActivity.class);
+                Intent intent = new Intent(ManageFacility.this, MainActivity.class);
                 startActivity(intent);
             }
         });
